✨ Feat: Footer 컴포넌트 - 링크 상수작업
